move range_str to utils, fix so it returns list of str
