+++ cnn_bilstm/utils.py
@@ -896,3 +896,50 @@ def convert_timebins_to_labels(labeled_timebins,
             return ''.join(labels)
         elif all([type(el) is int for el in labels]):
             return labels
+
+
+def range_str(range_str, sort=True):
+    """Generate range of ints from a formatted string,
+    then convert range from int to str
+
+    Example:
+        >>> range_str('1-4,6,9-11')
+        ['1','2','3','4','6','9','10','11']
+
+    Takes a range in form of "a-b" and returns
+    a list of numbers between a and b inclusive.
+    Also accepts comma separated ranges like "a-b,c-d,f"  which will
+    return a list with numbers from a to b, c to d, and f.
+
+    Parameters
+    ----------
+    range_str : str
+        of form 'a-b,c'
+        where a hyphen indicates a range
+        and a comma separates ranges or single numbers
+    sort : bool
+        If True, sort output before returning. Default is True.
+
+    Returns
+    -------
+    list_range : list
+        of int, produced by parsing range_str
+    """
+    # adapted from
+    # http://code.activestate.com/recipes/577279-generate-list-of-numbers-from-hyphenated-and-comma/
+    s = "".join(range_str.split())  # removes white space
+    list_range = []
+    for substr in range_str.split(','):
+        subrange = substr.split('-')
+        if len(subrange) not in [1, 2]:
+            raise SyntaxError("unable to parse range {} in labelset {}."
+                              .format(subrange, substr))
+        list_range.extend(
+            [int(subrange[0])]
+        ) if len(subrange) == 1 else list_range.extend(
+            range(int(subrange[0]), int(subrange[1]) + 1))
+
+    if sort:
+        list_range.sort()
+
+    return [str(list_int) for list_int in list_range]
