+++ make_data.py
@@ -5,53 +5,7 @@ from datetime import datetime
 import logging
 from configparser import ConfigParser
 
-from cnn_bilstm.utils import make_spects_from_list_of_files, make_data_dicts
-
-
-def range_str(range_str, sort=True):
-    """Generate range of ints from string
-
-    Example:
-        >>> range_str('1-4,6,9-11')
-        [1,2,3,4,6,9,10,11]
-
-    Takes a range in form of "a-b" and returns
-    a list of numbers between a and b inclusive.
-    Also accepts comma separated ranges like "a-b,c-d,f"  which will
-    return a list with numbers from a to b, c to d, and f.
-
-    Parameters
-    ----------
-    range_str : str
-        of form 'a-b,c'
-        where a hyphen indicates a range
-        and a comma separates ranges or single numbers
-    sort : bool
-        If True, sort output before returning. Default is True.
-
-    Returns
-    -------
-    list_range : list
-        of int, produced by parsing range_str
-    """
-    # adapted from
-    # http://code.activestate.com/recipes/577279-generate-list-of-numbers-from-hyphenated-and-comma/
-    s = "".join(range_str.split())  # removes white space
-    list_range = []
-    for substr in range_str.split(','):
-        subrange = substr.split('-')
-        if len(subrange) not in [1, 2]:
-            raise SyntaxError("unable to parse range {} in labelset {}."
-                              .format(subrange, substr))
-        list_range.extend(
-            [int(subrange[0])]
-        ) if len(subrange) == 1 else list_range.extend(
-            range(int(subrange[0]), int(subrange[1]) + 1))
-
-    if sort:
-        list_range.sort()
-
-    return list_range
+from cnn_bilstm.utils import make_spects_from_list_of_files, make_data_dicts, range_str
 
 
 if __name__ == "__main__":
